test: isolate DoorDash checkout binding guard
