+++ packages/cloud/shared/src/lib/services/doordash-browser-run.ts
@@ -4,10 +4,10 @@
  * inside Browser Run, while Live View provides an explicit human login handoff.
  */
 
-import { createHash } from "node:crypto";
 import type { Browser, BrowserWorker, Page } from "@cloudflare/playwright";
 import { getCloudBinding } from "../runtime/cloud-bindings";
 import { logger } from "../utils/logger";
+import { assertManagedCheckoutBinding } from "./doordash-checkout-binding";
 
 const BASE_URL = "https://www.doordash.com";
 const KEEP_ALIVE_MS = 600_000;
@@ -29,41 +29,6 @@ interface DoorDashHumanHandoff {
   readonly humanInterventionRequired: true;
   readonly handoffId?: string;
   readonly handoffState: "active" | "manual";
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return Boolean(value) && typeof value === "object" && !Array.isArray(value);
-}
-
-function stableValue(value: unknown): unknown {
-  if (Array.isArray(value)) return value.map(stableValue);
-  if (!isRecord(value)) return value;
-  return Object.fromEntries(
-    Object.entries(value)
-      .sort(([left], [right]) => left.localeCompare(right))
-      .map(([key, entry]) => [key, stableValue(entry)]),
-  );
-}
-
-export function managedCheckoutBindingDigest(cart: unknown, preview: unknown): string {
-  return createHash("sha256")
-    .update(JSON.stringify(stableValue({ cart, checkout: preview })))
-    .digest("hex");
-}
-
-export function assertManagedCheckoutBinding(
-  expectedDigest: unknown,
-  cart: unknown,
-  preview: unknown,
-): void {
-  if (typeof expectedDigest !== "string" || !/^[a-f0-9]{64}$/.test(expectedDigest)) {
-    throw new Error("DoorDash checkout requires the exact user-confirmed checkout digest");
-  }
-  if (managedCheckoutBindingDigest(cart, preview) !== expectedDigest) {
-    throw new Error(
-      "DoorDash cart or checkout changed after confirmation; review the new preview before ordering",
-    );
-  }
 }
 
 function browserBinding(): BrowserWorker {
